Fixed array problems, qsort works!!
There was a problem on recognizing variable declaration and after constant declaration inside functions. Solved the problem by adding another rule.
Also, the other problem was that the data type of a function declaration was mandatory(it shouldn't, and in case there is no data type, it is translated to void).